Refactor to split up the models completely
