+++ src/semantle_slack_bot/exceptions.py
@@ -2,6 +2,10 @@ class SemantleException(Exception):
     pass
 
 
+class InvalidWord(SemantleException):
+    pass
+
+
 class DatabaseException(SemantleException):
     pass
