+++ src/semantle_slack_bot/models/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa: F401
+from .game import Game
+from .guess import Guess
+from .nearby import Nearby
+from .similarity_range import SimilarityRange
+from .user import User
+from .word2vec import Word2Vec
