+++ src/semantle_slack_bot/slack.py
@@ -3,8 +3,8 @@ from typing import Literal, Optional, TypedDict
 
 from sqlalchemy.ext.asyncio.session import AsyncSession
 
-from semantle_slack_bot import db
 from semantle_slack_bot.logging import logger
+from semantle_slack_bot.models import Game, Guess
 from semantle_slack_bot.utils import get_custom_progress_bar
 
 SPACE = " "
@@ -69,11 +69,11 @@ class GuessContextBlock(TypedDict):
 class SlackGame:
     """A slack game instance"""
 
-    _game: db.Game
+    _game: Game
     input_action_id: str = "submit-guess"
     input_block_id: str = "guess-input"
 
-    def __init__(self, game: db.Game) -> None:
+    def __init__(self, game: Game) -> None:
         self._game = game
 
     @property
@@ -143,7 +143,7 @@ class SlackGame:
 
         return self.markdown_section(text=text)
 
-    def guess_context(self, guess: db.Guess, base_id: str) -> GuessContextBlock:
+    def guess_context(self, guess: Guess, base_id: str) -> GuessContextBlock:
         closeness = _closeness(guess)
 
         if guess.percentile == 1000:
@@ -170,7 +170,7 @@ class SlackGame:
         }
 
 
-def _closeness(guess: db.Guess) -> str:
+def _closeness(guess: Guess) -> str:
     if guess.percentile:
         if guess.percentile < 10:
             percentile = f"{SPACE * 7}{guess.percentile}"
@@ -187,7 +187,7 @@ def _closeness(guess: db.Guess) -> str:
     return f"{get_custom_progress_bar(0, 1000, width=6)}{SPACE * 14}cold"
 
 
-def _idx(guess: db.Guess) -> str:
+def _idx(guess: Guess) -> str:
     # Magic!
     # Add 6 spaces for idx < 10, 4 spaces for idx < 100, else 2 spaces
     # TODO: over 100 seems to not work
@@ -196,7 +196,7 @@ def _idx(guess: db.Guess) -> str:
     return f"{guess.idx}.{postfix}"
 
 
-def _similarity(guess: db.Guess) -> str:
+def _similarity(guess: Guess) -> str:
     prefix = ""
     if guess.similarity >= 0:
         # Account for negative symbol
@@ -208,5 +208,5 @@ def _similarity(guess: db.Guess) -> str:
     return f"_{prefix}{guess.similarity:.02f}_       "
 
 
-def _word(guess: db.Guess) -> str:
+def _word(guess: Guess) -> str:
     return f"*{guess.word}*"
